REFACTOR: Move post-send warnings into their own component.
Unfortunately, I couldn't refactor as much as I would like here. My original idea was the component to handle message bus subscriptions and callbacks to isolate post-send warnings from the `chat-live-pane` and `chat-message`. However, I couldn't make it work because we can't set the message ID inmediately after sending a message, meaning we won't be able to subscribe to the channel in time and miss the update. For this reason, we can't break the current design where the pane acts as a message broker, which dispatches updates to specific messages.

The change accomplished two things. It simplifies the warning MB message homogenizing all the different warning types into a single array, and renders a component that know how to translate them without needing multiple functions like the old implementation.